Money amount Int -> Long

+++ phoenix-scala/phoenix/app/phoenix/failures/CaptureFailures.scala
@@ -15,7 +15,7 @@ object CaptureFailures {
       s"The SKU $sku is missing a price"
   }
 
-  case class ShippingCostNegative(total: Int) extends Failure {
+  case class ShippingCostNegative(total: Long) extends Failure {
     override def description =
       s"Expected a shipping cost greater than zero but got $total"
   }
